Consolidate CLI error handling into shared errors module
Create src/magpie/cli/errors.py as the dedicated module for CLI error
handling, moving the unified error handler and token masking utilities
from utils.py. This addresses issue #70 by ensuring all CLI commands
use a single shared implementation.

Changes:
- Create errors.py with handle_http_error, format_auth_error, and
  mask_token functions
- Update all 11 CLI command files to import from errors module
- Remove duplicate _handle_error() from url.py (last remaining
  duplicated implementation)
- Update utils.py to re-export for backward compatibility
- Add backward compatibility tests to verify re-exports work

The url.py command was the only file still using a local _handle_error
function. All other commands had already been updated to use the shared
handler, but were importing from utils.py instead of the new errors.py.

Fixes #70
