+++ src/magpie/cli/commands/gc.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import click
 
 from magpie.cli import CLIContext
-from magpie.cli.utils import handle_http_error, mask_token
+from magpie.cli.errors import handle_http_error, mask_token
 
 
 @click.command()
