+++ src/magpie/cli/commands/get.py
@@ -9,8 +9,8 @@ import click
 
 from magpie.cli import CLIContext
 from magpie.cli.commands.parse import ParseError, parse_artifact_ref
+from magpie.cli.errors import handle_http_error
 from magpie.cli.progress import transfer_progress
-from magpie.cli.utils import handle_http_error
 from magpie.storage.hash import compute_hash
