+++ src/magpie/cli/commands/ls.py
@@ -11,7 +11,7 @@ if TYPE_CHECKING:
 
 from magpie.cli import CLIContext
 from magpie.cli.commands.parse import ParseError, parse_artifact_path
-from magpie.cli.utils import handle_http_error
+from magpie.cli.errors import handle_http_error
 
 
 @click.command(name="ls")
