+++ src/magpie/cli/commands/url.py
@@ -2,15 +2,11 @@
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
-
 import click
-
-if TYPE_CHECKING:
-    import httpx
 
 from magpie.cli import CLIContext
 from magpie.cli.commands.parse import ParseError, parse_artifact_ref
+from magpie.cli.errors import handle_http_error
 
 
 @click.command()
@@ -50,7 +46,7 @@ def url(ctx: CLIContext, artifact_ref: str) -> None:
         if response.status_code == 404:
             raise click.ClickException(f"Artifact not found: {parsed.path}:{parsed.ref}")
         if response.status_code != 200:
-            _handle_error(response)
+            handle_http_error(response, "URL resolution")
 
         data = response.json()
         hash_ref = data["hash_ref"]
@@ -65,13 +61,3 @@ def url(ctx: CLIContext, artifact_ref: str) -> None:
         # User requested tag - use tag symlink path
         download_url = f"{ctx.server}/artifacts/{parsed.path}/{parsed.ref}"
     click.echo(download_url)
-
-
-def _handle_error(response: "httpx.Response") -> None:
-    """Handle HTTP error responses."""
-    try:
-        detail = response.json().get("detail", response.text)
-    except Exception:
-        detail = response.text
-
-    raise click.ClickException(f"URL resolution failed ({response.status_code}): {detail}")
